Refactored logic into useRestoreScrollPosition hook
